Swapped out JSX for React icons with createElement on home and about pages
